Added a new class to help the sorted dictionary compare our objects and prevent duplicacy.
We also changed the hashset to a sorted set to allow us to sort the employees in the dictionary aplabetically both by department and by name.

To implement the sorting by name we added the IComparer interface to the Employee Comparer class in order to tell the SortedSet how to arrange employees.